requirements.txt
neo4j==5.14.1
langchain==0.1.0
langchain-community==0.0.16
langchain-core==0.1.17
pydantic==2.5.2
openai==1.12.0
python-dotenv==1.0.1
tiktoken==0.5.2
typing-inspect==0.9.0
langchain-openai==0.0.2 